pyproject.toml
[tool.poetry]
name = "polybot"
version = "0.1.0"
description = "Play polybot with your friends in Telegram chats"
authors = ["Milinuri Nirvalen <pentergust@mail.ru>"]
license = "GPL-3.0-or-later"
readme = "README.md"


# Dependencies ---------------------------------------------------------

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^11.0.0"
aiogram = "^3.15.0"
loguru = "^0.7.2"
tortoise-orm = "^0.22.1"


# Development dependencies ---------------------------------------------

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.1"
icecream = "^2.1.3"


# Ruff linter ----------------------------------------------------------

[tool.ruff]
line-length = 80
indent-width = 4

# Assume Python 3.10
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "D", "N", "PL"]
ignore = ["D407", "D107", "D213", "D203"]


# Configure build ------------------------------------------------------

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
